PlumbingProjectApp/backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
CacheControl==0.14.3
cachetools==6.2.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
filelock==3.20.0
firebase_admin==7.1.0
fireo==2.2.2
Flask==3.1.2
flask-cors==6.0.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.10.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-storage==3.5.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.72.0
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.62.3
gunicorn==22.0.0
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.0.5
langchain-classic==1.0.0
langchain-community==0.4.1
langchain-core==1.0.4
langchain-huggingface==1.0.1
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.42
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.8
msgpack==1.1.2
multidict==6.7.0
mypy_extensions==1.1.0
numpy==2.3.4
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==3.0.0
pillow==12.0.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==7.1.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tenacity==9.1.2
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0